♻️ Refactor tools API
