+++ plugins/webapi/tools/message/missing/handler.go
@@ -1,0 +1,25 @@
+package missing
+
+import (
+	"net/http"
+
+	"github.com/iotaledger/goshimmer/plugins/messagelayer"
+	"github.com/labstack/echo"
+)
+
+// Handler process missing requests.
+func Handler(c echo.Context) error {
+	res := &MissingResponse{}
+	missingIDs := messagelayer.Tangle().MissingMessages()
+	for _, msg := range missingIDs {
+		res.IDs = append(res.IDs, msg.String())
+	}
+	res.Count = len(missingIDs)
+	return c.JSON(http.StatusOK, res)
+}
+
+// MissingResponse is the HTTP response containing all the missing messages and their count.
+type MissingResponse struct {
+	IDs   []string `json:"ids,omitempty"`
+	Count int      `json:"count,omitempty"`
+}
