+++ plugins/webapi/tools/plugin.go
@@ -1,17 +1,15 @@
 package tools
 
 import (
-	"container/list"
-	"fmt"
-	"net/http"
 	"sync"
 
-	"github.com/iotaledger/goshimmer/packages/binary/messagelayer/message"
-	"github.com/iotaledger/goshimmer/plugins/messagelayer"
 	"github.com/iotaledger/goshimmer/plugins/webapi"
+	"github.com/iotaledger/goshimmer/plugins/webapi/tools/message/missing"
+	"github.com/iotaledger/goshimmer/plugins/webapi/tools/message/pastcone"
+	"github.com/iotaledger/goshimmer/plugins/webapi/tools/value/debug"
+	"github.com/iotaledger/goshimmer/plugins/webapi/tools/value/tips"
 	"github.com/iotaledger/hive.go/logger"
 	"github.com/iotaledger/hive.go/node"
-	"github.com/labstack/echo"
 )
 
 // PluginName is the name of the web API tools endpoint plugin.
@@ -34,100 +32,8 @@ func Plugin() *node.Plugin {
 
 func configure(_ *node.Plugin) {
 	log = logger.NewLogger(PluginName)
-	webapi.Server().GET("tools/pastcone", pastCone)
-	webapi.Server().GET("tools/missing", missing)
-}
-
-func pastCone(c echo.Context) error {
-	var checkedMessageCount int
-	var request PastConeRequest
-	if err := c.Bind(&request); err != nil {
-		log.Info(err.Error())
-		return c.JSON(http.StatusBadRequest, PastConeResponse{Error: err.Error()})
-	}
-
-	log.Info("Received:", request.ID)
-
-	msgID, err := message.NewID(request.ID)
-	if err != nil {
-		log.Info(err)
-		return c.JSON(http.StatusBadRequest, PastConeResponse{Error: err.Error()})
-	}
-
-	// create a new stack that hold messages to check
-	stack := list.New()
-	stack.PushBack(msgID)
-	// keep track of submitted checks (to not re-add something to the stack that is already in it)
-	// searching in double-linked list is quite expensive, but not in a map
-	submitted := make(map[message.ID]bool)
-
-	// process messages in stack, try to request parents until we end up at the genesis
-	for stack.Len() > 0 {
-		checkedMessageCount++
-		// pop the first element from stack
-		currentMsgElement := stack.Front()
-		currentMsgID := currentMsgElement.Value.(message.ID)
-		stack.Remove(currentMsgElement)
-
-		// ask node if it has it
-		msgObject := messagelayer.Tangle().Message(currentMsgID)
-		msgMetadataObject := messagelayer.Tangle().MessageMetadata(currentMsgID)
-
-		if !msgObject.Exists() || !msgMetadataObject.Exists() {
-			return c.JSON(http.StatusOK, PastConeResponse{Exist: false, PastConeSize: checkedMessageCount, Error: fmt.Sprintf("couldn't find %s message on node", currentMsgID)})
-		}
-
-		// get trunk and branch
-		msg := msgObject.Unwrap()
-		branchID := msg.BranchID()
-		trunkID := msg.TrunkID()
-
-		// release objects
-		msgObject.Release()
-		msgMetadataObject.Release()
-
-		if branchID == message.EmptyID && msg.TrunkID() == message.EmptyID {
-			// msg only attaches to genesis
-			continue
-		} else {
-			if !submitted[branchID] && branchID != message.EmptyID {
-				stack.PushBack(branchID)
-				submitted[branchID] = true
-			}
-			if !submitted[trunkID] && trunkID != message.EmptyID {
-				stack.PushBack(trunkID)
-				submitted[trunkID] = true
-			}
-		}
-	}
-	return c.JSON(http.StatusOK, PastConeResponse{Exist: true, PastConeSize: checkedMessageCount})
-}
-
-// PastConeRequest holds the message id to query.
-type PastConeRequest struct {
-	ID string `json:"id"`
-}
-
-// PastConeResponse is the HTTP response containing the number of messages in the past cone and if all messages of the past cone
-// exist on the node.
-type PastConeResponse struct {
-	Exist        bool   `json:"exist,omitempty"`
-	PastConeSize int    `json:"pastConeSize,omitempty"`
-	Error        string `json:"error,omitempty"`
-}
-
-func missing(c echo.Context) error {
-	res := &MissingResponse{}
-	missingIDs := messagelayer.Tangle().MissingMessages()
-	for _, msg := range missingIDs {
-		res.IDs = append(res.IDs, msg.String())
-	}
-	res.Count = len(missingIDs)
-	return c.JSON(http.StatusOK, res)
-}
-
-// MissingResponse is the HTTP response containing all the missing messages and their count.
-type MissingResponse struct {
-	IDs   []string `json:"ids,omitempty"`
-	Count int      `json:"count,omitempty"`
+	webapi.Server().GET("tools/message/pastcone", pastcone.Handler)
+	webapi.Server().GET("tools/message/missing", missing.Handler)
+	webapi.Server().GET("tools/value/tips", tips.Handler)
+	webapi.Server().GET("tools/value/debug", debug.Handler)
 }
