+++ plugins/webapi/value/plugin.go
@@ -5,12 +5,10 @@ import (
 
 	"github.com/iotaledger/goshimmer/plugins/webapi"
 	"github.com/iotaledger/goshimmer/plugins/webapi/value/attachments"
-	"github.com/iotaledger/goshimmer/plugins/webapi/value/debug"
 	"github.com/iotaledger/goshimmer/plugins/webapi/value/gettransactionbyid"
 	"github.com/iotaledger/goshimmer/plugins/webapi/value/sendtransaction"
 	"github.com/iotaledger/goshimmer/plugins/webapi/value/sendtransactionbyjson"
 	"github.com/iotaledger/goshimmer/plugins/webapi/value/testsendtxn"
-	"github.com/iotaledger/goshimmer/plugins/webapi/value/tips"
 	"github.com/iotaledger/goshimmer/plugins/webapi/value/unspentoutputs"
 	"github.com/iotaledger/hive.go/node"
 )
@@ -39,6 +37,4 @@ func configure(_ *node.Plugin) {
 	webapi.Server().POST("value/sendTransactionByJson", sendtransactionbyjson.Handler)
 	webapi.Server().POST("value/testSendTxn", testsendtxn.Handler)
 	webapi.Server().GET("value/transactionByID", gettransactionbyid.Handler)
-	webapi.Server().GET("value/tips", tips.Handler)
-	webapi.Server().GET("value/debug", debug.Handler)
 }
